requirements.txt
sentence-transformers
langchain-community
langchain
langchain_core
discord.py
python-dotenv
chromadb
deepeval